requirements.txt
# requirements.txt

python-dotenv
groq
pandas
numpy
python-dotenv
pyyaml
flask
markdown
dotenv
faiss-cpu
sentence-transformers
PyPDF2

